Fixs orders_service owner access

+++ orders_service/app/routers/orders_crud.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.db_depends import get_db
@@ -12,7 +12,7 @@ from app.crud.orders import (
     update_order_status_in_db,
     delete_order_from_db,
 )
-from app.dependencies.depend import authentication_get_current_user, permission_required, user_owner_access_checker
+from app.dependencies.depend import authentication_get_current_user, permission_required #, user_owner_access_checker
 
 
 router = APIRouter(prefix="/orders_crud", tags=["Orders CRUD"])
@@ -23,9 +23,16 @@ router = APIRouter(prefix="/orders_crud", tags=["Orders CRUD"])
             response_model=list[OrderOut])
 async def get_all_orders(
     db: AsyncSession = Depends(get_db),
-    user=Depends(authentication_get_current_user),
+    current_user=Depends(authentication_get_current_user),
 ):
-    user_id = user["id"]
+    perms = current_user.get("permissions") or []
+    if "can_force_get_order" not in perms:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not allowed to access, for get all orders",
+        )
+
+    user_id = current_user["id"]
     if isinstance(user_id, str):
         user_id = UUID(user_id)
     return await get_all_orders_from_db(db, user_id)
@@ -37,10 +44,20 @@ async def get_all_orders(
 async def get_order(
     order_id: UUID,
     db: AsyncSession = Depends(get_db),
-    user=Depends(authentication_get_current_user),
-    _: None = Depends(user_owner_access_checker),
+    current_user=Depends(authentication_get_current_user),
 ):
-    return await get_order_from_db(order_id, db)
+    order = await get_order_from_db(db, order_id)
+    if not order:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
+
+    user_id = UUID(current_user["id"])
+    perms = current_user.get("permissions") or []
+    if (order.user_id != user_id) and ("can_force_get_order" not in perms):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not allowed to access this order (owner only)",
+        )
+    return order
 
 
 @router.post("/",
@@ -65,9 +82,19 @@ async def update_order_status(
     order_id: UUID,
     data: OrderStatusPatch,
     db: AsyncSession = Depends(get_db),
-    user=Depends(authentication_get_current_user),
-    _: None = Depends(user_owner_access_checker),
+    current_user=Depends(authentication_get_current_user),
 ):
+    order = await get_order_from_db(db, order_id)
+    if not order:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
+
+    user_id = UUID(current_user["id"])
+    perms = current_user.get("permissions") or []
+    if (order.user_id != user_id) and ("can_force_get_order" not in perms):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not allowed to access this order (owner only)",
+        )
     return await update_order_status_in_db(order_id, data, db)
 
 
@@ -76,7 +103,17 @@ async def update_order_status(
 async def delete_order(
     order_id: UUID,
     db: AsyncSession = Depends(get_db),
-    user=Depends(authentication_get_current_user),
-    _: None = Depends(user_owner_access_checker),
+    current_user=Depends(authentication_get_current_user),
 ):
+    order = await get_order_from_db(db, order_id)
+    if not order:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
+
+    user_id = UUID(current_user["id"])
+    perms = current_user.get("permissions") or []
+    if (order.user_id != user_id) and ("can_force_get_order" not in perms):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not allowed to access this order (owner only)",
+        )
     return await delete_order_from_db(order_id, db)
